fix(l3gd20h): fix CI failures and hallucinated test/API mismatches across all languages
- Rust: fix wrong path deps in example/test Cargo.tomls, rename colliding
  gyroscope const re-exports and helper fns (l3g4200d/l3gd20h collision),
  drop erroneous Connection::new() wrapping (driver takes the raw bus).
- Go: add missing //go:build tinygo tags, fix TinyGo NewI2CConnection call
  shape, fix I2C addr type mismatch, rename colliding sensitivity/int16Le
  helpers and mockConn type, fix mock WriteRead register masking.
- JVM: rewrite Java/Kotlin/Groovy L3gd20h unit tests, which called a
  fictional MockConnection builder API and L3G4200D-shaped methods
  (setDuration, readIntSource, etc.) that L3GD20H's spec never defines;
  fix real Kotlin/Groovy compile errors in the production driver
  (stray @JvmOverloads, val reassignment, private companion members,
  missing imports, float precision).
- Python/Node.js/C++: rewrite unit tests that referenced nonexistent mock
  APIs (I2CMock/set_reg/get_reg, expectWrite/verify) instead of the real
  I2CConnectionMock/I2CConnectionMock register-map mocks; fix a missing
  0x80 auto-increment bit in the Node.js mock register setup.
- C++: add missing Arduino delay_ms() wrapper (only Zephyr/ESP-IDF/Pico
  SDK/Linux had one), which broke Arduino compilation.
- Regenerate stale cpp/README.md and nodejs package READMEs.

+++ go/examples/linux/gyroscope/l3gd20h/minimal/minimal.go
@@ -18,7 +18,7 @@ func main() {
 		fmt.Sscanf(a, "%x", &addr)
 	}
 
-	conn, err := connection.NewI2CConnection(bus, addr, nil, nil)
+	conn, err := connection.NewI2CConnection(bus, uint8(addr), nil, nil)
 	if err != nil {
 		panic(err)
 	}
